Move value package to parser/json
